Moved controlchan::commands::handler to controlchan::handler
