+++ src/server/controlchan/commands/acct.rs
@@ -17,8 +17,8 @@
 // (pending receipt of the ACCounT command) or discards the
 // command, respectively.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
