+++ src/server/controlchan/commands/allo.rs
@@ -18,8 +18,8 @@
 // the maximum record or page size should accept a dummy value
 // in the first argument and ignore it.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
