+++ src/server/controlchan/commands/auth.rs
@@ -4,9 +4,9 @@
 //! wishes to secure the data connections by use of the PBSZ and PROT
 //! commands.
 
-use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
