+++ src/server/controlchan/commands/ccc.rs
@@ -1,8 +1,8 @@
 //! The RFC 2228 Clear Command Channel (`CCC`) command
 
-use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
