+++ src/server/controlchan/commands/cdup.rs
@@ -6,8 +6,8 @@
 // syntaxes for naming the parent directory.  The reply codes
 // shall be identical to the reply codes of CWD.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
