+++ src/server/controlchan/commands/feat.rs
@@ -1,7 +1,7 @@
 //! The RFC 2389 Feature (`FEAT`) command
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
