+++ src/server/controlchan/commands/mod.rs
@@ -13,7 +13,6 @@ mod cdup;
 mod cwd;
 mod dele;
 mod feat;
-mod handler;
 mod help;
 mod list;
 mod mdtm;
@@ -52,7 +51,6 @@ pub use cdup::Cdup;
 pub use cwd::Cwd;
 pub use dele::Dele;
 pub use feat::Feat;
-pub(crate) use handler::{CommandContext, CommandHandler};
 pub use help::Help;
 pub use list::List;
 pub use mdtm::Mdtm;
