+++ src/server/controlchan/commands/nlst.rs
@@ -13,9 +13,9 @@
 // further process the files automatically.  For example, in
 // the implementation of a "multiple get" function.
 
-use super::handler::CommandContext;
 use crate::server::controlchan::command::Command;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
