+++ src/server/controlchan/commands/noop.rs
@@ -4,8 +4,8 @@
 // entered commands. It specifies no action other than that the
 // server send an OK reply.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
