+++ src/server/controlchan/commands/opts.rs
@@ -7,8 +7,8 @@
 // definition of that command.  Where no OPTS behavior is defined for a
 // particular command there are no options available for that command.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
