+++ src/server/controlchan/commands/pass.rs
@@ -10,9 +10,9 @@
 // therefore the responsibility of the user-FTP process to hide
 // the sensitive password information.
 
-use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::server::password;
