+++ src/server/controlchan/commands/pasv.rs
@@ -6,8 +6,8 @@
 // transfer command.  The response to this command includes the
 // host and port address this server is listening on.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::Command;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
