+++ src/server/controlchan/commands/pbsz.rs
@@ -9,8 +9,8 @@
 //! of '0' to indicate that no buffering is taking place and the data connection should
 //! not be encapsulated.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
