+++ src/server/controlchan/commands/prot.rs
@@ -1,7 +1,7 @@
 //! The RFC 2228 Data Channel Protection Level (`PROT`) command.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
