+++ src/server/controlchan/commands/pwd.rs
@@ -3,8 +3,8 @@
 // This command causes the name of the current working
 // directory to be returned in the reply.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
