+++ src/server/controlchan/commands/rmd.rs
@@ -5,9 +5,9 @@
 // or as a subdirectory of the current working directory (if
 // the pathname is relative).
 
-use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::Reply;
 use crate::server::error::FTPError;
 use crate::storage;
