+++ src/server/controlchan/commands/stat.rs
@@ -17,9 +17,9 @@
 // should include current values of all transfer parameters and
 // the status of connections.
 
-use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage::{self, Error, ErrorKind};
