+++ src/server/controlchan/commands/stor.rs
@@ -8,9 +8,9 @@
 // created at the server site if the file specified in the
 // pathname does not already exist.
 
-use super::handler::CommandContext;
 use crate::server::controlchan::command::Command;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
