+++ src/server/controlchan/commands/stou.rs
@@ -1,8 +1,8 @@
 //! The RFC 959 Store File Uniquely (`STOU`) command
 
-use super::handler::CommandContext;
 use crate::server::controlchan::command::Command;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
