+++ src/server/controlchan/commands/type_.rs
@@ -26,8 +26,8 @@
 // argument is changed, Format then returns to the Non-print
 // default.
 
-use super::handler::CommandContext;
-use crate::server::controlchan::commands::CommandHandler;
+use crate::server::controlchan::handler::CommandContext;
+use crate::server::controlchan::handler::CommandHandler;
 use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::storage;
